fix(ListNodes): handled node images over 2GB size

+++ .gen/pipeline/pipeline/model_node_item_status_images.go
@@ -14,5 +14,5 @@ type NodeItemStatusImages struct {
 
 	Name []string `json:"name,omitempty"`
 
-	SizeBytes int32 `json:"sizeBytes,omitempty"`
+	SizeBytes int64 `json:"sizeBytes,omitempty"`
 }
